requirements.txt
crewai
pandas
python-dotenv
crewai_tools